+++ Stripe/AfterpayPriceBreakdownView.swift
@@ -10,18 +10,10 @@ import UIKit
 import SafariServices
 @_spi(STP) import StripeUICore
 
-/// The view looks like:
-///
-/// Single row (width can contain all subviews):
-///   Pay in 4 interest-free payments of %@ with [Afterpay logo] [info button]
-///
-/// Multi row (width can't contain all subviews):
-///   Pay in 4 interest-free payments of %@ with
-///   [Afterpay logo] [info button]
 /// For internal SDK use only
 @objc(STP_Internal_AfterpayPriceBreakdownView)
 class AfterpayPriceBreakdownView: UIView {
-    private let priceBreakdownLabel = UILabel()
+    private let afterPayClearPayLabel = UILabel()
     
     private lazy var afterpayMarkImageView: UIImageView = {
         let imageView = UIImageView()
@@ -31,50 +23,48 @@ class AfterpayPriceBreakdownView: UIView {
 
         return imageView
     }()
-    
-    private lazy var infoButton: UIButton = {
-        let button = UIButton()
-        button.setImage(STPImageLibrary.safeImageNamed("afterpay_icon_info@3x"), for: .normal)
-        return button
+    private lazy var afterpayMarkImage: UIImage = {
+        return STPImageLibrary.afterpayLogo(locale: locale)
     }()
-    
+    private lazy var infoImage: UIImage = {
+        return STPImageLibrary.safeImageNamed("afterpay_icon_info@3x")
+    }()
+
     private lazy var infoURL: URL? = {
         let regionCode = Locale.current.regionCode ?? "us"
         return URL(string: "https://static-us.afterpay.com/javascript/modal/\(regionCode.lowercased())_rebrand_modal.html")
     }()
+
+    static func numberOfInstallments(currency: String) -> Int {
+        return currency.uppercased() == "EUR" ? 3 : 4
+    }
 
     let locale: Locale
     
     init(amount: Int, currency: String, locale: Locale = Locale.autoupdatingCurrent) {
         self.locale = locale
         super.init(frame: .zero)
-        
-        let installmentAmount = amount / 4
+        let numInstallments = Self.numberOfInstallments(currency: currency)
+        let installmentAmount = amount / numInstallments
         let installmentAmountDisplayString = String.localizedAmountDisplayString(for: installmentAmount, currency: currency)
         
-        priceBreakdownLabel.attributedText = generatePriceBreakdownString(installmentAmountString: installmentAmountDisplayString)
-        
-        [priceBreakdownLabel, afterpayMarkImageView, infoButton].forEach {
-            $0.translatesAutoresizingMaskIntoConstraints = false
-            addSubview($0)
-        }
+        afterPayClearPayLabel.attributedText = generateAfterPayClearPayString(numInstallments: numInstallments,
+                                                                              installmentAmountString: installmentAmountDisplayString)
+        afterPayClearPayLabel.numberOfLines = 0
+        afterPayClearPayLabel.translatesAutoresizingMaskIntoConstraints = false
+        afterPayClearPayLabel.isUserInteractionEnabled = true
+        addSubview(afterPayClearPayLabel)
 
         NSLayoutConstraint.activate([
-            priceBreakdownLabel.leadingAnchor.constraint(
+            afterPayClearPayLabel.leadingAnchor.constraint(
                 equalTo: leadingAnchor),
-            priceBreakdownLabel.topAnchor.constraint(
+            afterPayClearPayLabel.topAnchor.constraint(
                 equalTo: topAnchor),
-
-            afterpayMarkImageView.bottomAnchor.constraint(
-                equalTo: bottomAnchor),
-
-            infoButton.leadingAnchor.constraint(
-                equalTo: afterpayMarkImageView.trailingAnchor, constant: 7),
-            infoButton.bottomAnchor.constraint(
-                equalTo: bottomAnchor),
+            afterPayClearPayLabel.bottomAnchor.constraint(equalTo: bottomAnchor),
+            afterPayClearPayLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
         ])
         
-        infoButton.addTarget(self, action: #selector(didTapInfoButton), for: .touchUpInside)
+        afterPayClearPayLabel.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(didTapInfoButton)))
     }
     
     override init(frame: CGRect) {
@@ -86,79 +76,84 @@ class AfterpayPriceBreakdownView: UIView {
         fatalError("init(coder:) has not been implemented")
     }
 
-    private lazy var singleRowConstraints: [NSLayoutConstraint] = [
-
-        afterpayMarkImageView.leadingAnchor.constraint(
-            equalTo: priceBreakdownLabel.trailingAnchor, constant: 5),
-        afterpayMarkImageView.centerYAnchor.constraint(
-            equalTo: priceBreakdownLabel.centerYAnchor),
-
-        infoButton.centerYAnchor.constraint(
-            equalTo: priceBreakdownLabel.centerYAnchor),
-    ]
-
-    private lazy var multiRowConstraints: [NSLayoutConstraint] = [
-
-        afterpayMarkImageView.leadingAnchor.constraint(
-            equalTo: leadingAnchor),
-        afterpayMarkImageView.topAnchor.constraint(
-            equalTo: priceBreakdownLabel.bottomAnchor, constant: 2),
-
-        infoButton.centerYAnchor.constraint(
-            equalTo: afterpayMarkImageView.centerYAnchor),
-    ]
-    
     override func layoutSubviews() {
         super.layoutSubviews()
-        
-        if !subviewsOutOfBounds() {
-            NSLayoutConstraint.deactivate(multiRowConstraints)
-            NSLayoutConstraint.activate(singleRowConstraints)
-            super.layoutSubviews()
-        } else {
-            NSLayoutConstraint.deactivate(singleRowConstraints)
-            NSLayoutConstraint.activate(multiRowConstraints)
-            super.layoutSubviews()
-        }
     }
-    
-    private func subviewsOutOfBounds() -> Bool {
-        let subviewsTotalWidth = [priceBreakdownLabel, afterpayMarkImageView, infoButton].reduce(0) { $0 + $1.bounds.width }
-        return subviewsTotalWidth >= bounds.width
-    }
-    
-    private func generatePriceBreakdownString(installmentAmountString: String) -> NSMutableAttributedString {
+
+    private func generateAfterPayClearPayString(numInstallments: Int, installmentAmountString: String) -> NSMutableAttributedString {
         let amountStringAttributes = [
             NSAttributedString.Key.font: ElementsUITheme.current.fonts.subheadlineBold,
             .foregroundColor: ElementsUITheme.current.colors.bodyText
         ]
-        
         let stringAttributes = [
             NSAttributedString.Key.font: ElementsUITheme.current.fonts.subheadline,
             .foregroundColor: ElementsUITheme.current.colors.bodyText
         ]
-        
-        let amountString = NSMutableAttributedString(
-            string: "\(installmentAmountString) ",
-            attributes: amountStringAttributes
-        )
+        let template = STPLocalizedString("Pay in <num_installments/> interest-free payments of <installment_price/> with <img/>",
+                                          "Pay in templated string for afterpay/clearpay")
 
-        let payIn4String = NSMutableAttributedString(
-            string: "Pay in 4 interest-free payments of ",
-            attributes: stringAttributes
-        )
-        
-        let withString = NSMutableAttributedString(
-            string: "with",
-            attributes: stringAttributes
-        )
-        
-        payIn4String.append(amountString)
-        payIn4String.append(withString)
-        
-        return payIn4String
+        let resultingString = NSMutableAttributedString()
+        resultingString.append(NSAttributedString(string: ""))
+        guard let numInstallmentsRange = template.range(of: "<num_installments/>"),
+              let installmentPrice = template.range(of: "<installment_price/>"),
+              let img = template.range(of: "<img/>") else {
+            return resultingString
+        }
+
+        var numInstallmentsAppended = false
+        var installmentPriceAppended = false
+        var imgAppended = false
+
+        for (indexOffset, currCharacter) in template.enumerated() {
+            let currIndex = template.index(template.startIndex, offsetBy: indexOffset)
+            if numInstallmentsRange.contains(currIndex) {
+                if numInstallmentsAppended {
+                    continue
+                }
+                numInstallmentsAppended = true
+                resultingString.append(NSAttributedString(string: "\(numInstallments)",
+                                                          attributes: stringAttributes))
+            } else if installmentPrice.contains(currIndex) {
+                if installmentPriceAppended {
+                    continue
+                }
+                installmentPriceAppended = true
+                resultingString.append(NSAttributedString(string: installmentAmountString,
+                                                          attributes: amountStringAttributes))
+            } else if img.contains(currIndex) {
+                if imgAppended {
+                    continue
+                }
+                imgAppended = true
+                let titleFont = stringAttributes[NSAttributedString.Key.font] as! UIFont
+                let clearPay = attributedStringOfImageWithoutLink(uiImage: afterpayMarkImage, font: titleFont)
+                let infoButton = attributedStringOfImageWithoutLink(uiImage: infoImage, font: titleFont)
+                resultingString.append(clearPay)
+                resultingString.append(NSAttributedString(string: "\u{00A0}\u{00A0}", attributes: stringAttributes))
+                resultingString.append(infoButton)
+            } else {
+                resultingString.append(NSAttributedString(string: String(currCharacter),
+                                                          attributes: stringAttributes))
+            }
+        }
+        return resultingString
     }
-    
+
+    private func attributedStringOfImageWithoutLink(uiImage: UIImage, font: UIFont) -> NSAttributedString {
+        let imageAttachment = NSTextAttachment()
+        imageAttachment.bounds = boundsOfImage(font: font, uiImage: uiImage)
+        imageAttachment.image = uiImage
+        return NSAttributedString(attachment: imageAttachment)
+    }
+
+    // https://stackoverflow.com/questions/26105803/center-nstextattachment-image-next-to-single-line-uilabel
+    private func boundsOfImage(font: UIFont, uiImage: UIImage) -> CGRect {
+        return CGRect(x: 0,
+                      y: (font.capHeight - uiImage.size.height).rounded() / 2,
+                      width: uiImage.size.width,
+                      height: uiImage.size.height)
+    }
+
     @objc
     private func didTapInfoButton() {
         if let url = infoURL {