+++ Stripe/STPImageLibrary.swift
@@ -108,8 +108,8 @@ public class STPImageLibrary: NSObject {
     /// An icon representing Afterpay.
     @objc
     public class func afterpayLogo(locale: Locale = Locale.current) -> UIImage {
-        switch (locale.languageCode, locale.regionCode) {
-            case ("en", "GB"):
+        switch (locale.regionCode) {
+            case "GB", "FR", "ES":
                 return self.safeImageNamed("clearpay_mark", templateIfAvailable: true)
             default:
                return self.safeImageNamed("afterpay_mark", templateIfAvailable: true)